Handle C++x11 Enum names from CastXML
In order to support as many compiler tools as possible (e.g. Travis)
this change fixes things to look for enum names the old and the new way.

The old way does not include the enum's name itself in the fully
qualified name, and the new way does, so if you have an enum like
`::default_arguments::Ops::ADD` (the new way) it should always show up
as `::default_arguments::ADD` (the old way).

The new way works great on my local machine (macOS High Sierra) but
fails on Travis, so good catch there!